Refactor nimba sms application with redesign the architecture
