+++ README.md
@@ -19,29 +19,29 @@ System.out.println(accountResponse);
 Retrieve the sender names using the getSenderName() method:
 
 ```java
-RootResult<SenderNameResponse> senderNameResponse = client.getSenderName().list();
+SenderNameResponse senderNameResponse = client.getSenderName().list();
 System.out.println(senderNameResponse);
 ```
 You can also retrieve the last 10 sender names by passing in the limit and offset:
 
 ```java
-RootResult<SenderNameResponse> last10SenderNameResponse = client.getSenderName().list(10, 1);
+SenderNameResponse last10SenderNameResponse = client.getSenderName().list(10, 1);
 System.out.println(last10SenderNameResponse);
 ```
 ## Contacts
 This code retrieves a list of all contacts.
 ```java
-RootResult<ContactResponse> contacts = client.getContact().list();
+ContactResponse contacts = client.getContact().list();
 System.out.println(contacts);
 ```
 You can also retrieve the last 10 contacts by passing in the limit and offset:
 ```java
-RootResult<ContactResponse> last10contacts = client.getContact().list(10, 1);
+ContactResponse last10contacts = client.getContact().list(10, 1);
 System.out.println(last10contacts);
 ```
 Create Contact. This contact will be added to the default contact list:
 ```java
-ContactResponse defaultContactResponse = client.getContact().create("+224XXXXXXXXX", null, null);
+ContactDto defaultContactResponse = client.getContact().create("+224XXXXXXXXX", null, null);
 System.out.println(defaultContactResponse)
 ```
 
@@ -49,24 +49,24 @@ System.out.println(defaultContactResponse)
 ## Groups
 This code retrieves a list of all groups.
 ```java
-RootResult<GroupResponse> groups = client.getGroup().list();
+GroupResponse groups = client.getGroup().list();
 System.out.println(groups);
 ```
 
 You can also retrieve the last 10 Group by passing in the limit and offset:
 ```java
-RootResult<GroupResponse> last10groups = client.getGroup().list(10, 1);
+GroupResponse last10groups = client.getGroup().list(10, 1);
 System.out.println(last10groups);
 ```
 ## Message
 Get All messages
 ```java
-RootResult<MessageResponse> messages = client.getMessage().list();
+MessageResponse messages = client.getMessage().list();
 System.out.println(messages);
 ```
 Get only last 10 messages
 ```java
-RootResult<MessageResponse> last10Messages = client.getMessage().list(10, 1);
+MessageResponse last10Messages = client.getMessage().list(10, 1);
 System.out.println(last10Messages);
 ```
 Send a Message
@@ -76,7 +76,7 @@ System.out.println(messageResponse);
 ```
 Retrieve message
 ```java
-MessageDetails messageDetails = client.getMessage().retrieve("123");
+MessageDetailsResponse messageDetails = client.getMessage().retrieve("123");
 System.out.println(messageDetails);
 ```
