+++ pom.xml
@@ -14,7 +14,8 @@
 		<maven.compiler.source>11</maven.compiler.source>
 		<maven.compiler.target>11</maven.compiler.target>
 		<maven.compiler.version>3.6.1</maven.compiler.version>
-		<lombok.version>1.18.20</lombok.version>
+		<lombok.version>1.18.26</lombok.version>
+		<jackson-databind.version>2.15.1</jackson-databind.version>
 	</properties>
 	<dependencies>
 		<dependency>
@@ -31,7 +32,7 @@
 		<dependency>
 			<groupId>com.fasterxml.jackson.core</groupId>
 			<artifactId>jackson-databind</artifactId>
-			<version>2.14.1</version>
+			<version>${jackson-databind.version}</version>
 		</dependency>
