+++ src/main/java/com/nimbasms/nimbasms/accounts/Link.java
@@ -1,0 +1,11 @@
+package com.nimbasms.nimbasms.accounts;
+
+import lombok.*;
+
+@Data
+@NoArgsConstructor
+public class Link {
+    private String desc;
+    private String href;
+    private String method;
+}
