+++ src/main/java/com/nimbasms/nimbasms/contacts/ContactResponse.java
@@ -1,0 +1,6 @@
+package com.nimbasms.nimbasms.contacts;
+
+import com.nimbasms.nimbasms.common.ApiResponse;
+
+public class ContactResponse extends ApiResponse<ContactDto> {
+}
