+++ src/main/java/com/nimbasms/nimbasms/exception/NimbaSmsException.java
@@ -1,0 +1,10 @@
+package com.nimbasms.nimbasms.exception;
+
+import lombok.Getter;
+
+@Getter
+public class NimbaSmsException extends RuntimeException {
+    public NimbaSmsException(String message) {
+        super(message);
+    }
+}
