+++ src/main/java/com/nimbasms/nimbasms/groups/GroupResponse.java
@@ -1,0 +1,6 @@
+package com.nimbasms.nimbasms.groups;
+
+import com.nimbasms.nimbasms.common.ApiResponse;
+
+public class GroupResponse extends ApiResponse<GroupDto> {
+}
