+++ src/main/java/com/nimbasms/nimbasms/messages/MessageResponse.java
@@ -1,0 +1,6 @@
+package com.nimbasms.nimbasms.messages;
+
+import com.nimbasms.nimbasms.common.ApiResponse;
+
+public class MessageResponse extends ApiResponse<MessageDto> {
+}
