+++ src/main/java/com/nimbasms/nimbasms/messages/Number.java
@@ -1,0 +1,14 @@
+package com.nimbasms.nimbasms.messages;
+
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+
+@Getter
+@Setter
+@ToString
+public class Number {
+    private String id;
+    private String contact;
+    private String status;
+}
