+++ src/main/java/com/nimbasms/nimbasms/sendernames/SenderNameResponse.java
@@ -1,0 +1,6 @@
+package com.nimbasms.nimbasms.sendernames;
+
+import com.nimbasms.nimbasms.common.ApiResponse;
+
+public class SenderNameResponse extends ApiResponse<SenderNameDto> {
+}
