--- src/main/java/gn/nimba/nimbasms/accounts/Account.java
@@ -1,7 +1,0 @@
-package gn.nimba.nimbasms.accounts;
-
-import java.io.IOException;
-
-public interface Account {
-    AccountResponse get() throws IOException;
-}
