--- src/main/java/gn/nimba/nimbasms/accounts/AccountResponse.java
@@ -1,15 +1,0 @@
-package gn.nimba.nimbasms.accounts;
-
-import lombok.*;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import java.util.List;
-
-@Data
-@NoArgsConstructor
-public class AccountResponse {
-    private String sid;
-    private Integer balance;
-
-    @JsonProperty("_links")
-    private List<Link> links;
-}
