--- src/main/java/gn/nimba/nimbasms/accounts/Link.java
@@ -1,11 +1,0 @@
-package gn.nimba.nimbasms.accounts;
-
-import lombok.*;
-
-@Data
-@NoArgsConstructor
-public class Link {
-    private String desc;
-    private String href;
-    private String method;
-}
