--- src/main/java/gn/nimba/nimbasms/common/UriType.java
@@ -1,19 +1,0 @@
-package gn.nimba.nimbasms.common;
-
-public enum UriType {
-    ACCOUNTS("accounts"),
-    CONTACTS("contacts"),
-    GROUPS("groups"),
-    SENDER_NAMES("sendernames"),
-    MESSAGES("messages");
-
-    private final String path;
-
-    UriType(String path) {
-        this.path = path;
-    }
-
-    public String getPath() {
-        return path;
-    }
-}
