--- src/main/java/gn/nimba/nimbasms/exception/NimbaSmsException.java
@@ -1,10 +1,0 @@
-package gn.nimba.nimbasms.exception;
-
-import lombok.Getter;
-
-@Getter
-public class NimbaSmsException extends RuntimeException {
-    public NimbaSmsException(String message) {
-        super(message);
-    }
-}
