--- src/main/java/gn/nimba/nimbasms/groups/Group.java
@@ -1,6 +1,0 @@
-package gn.nimba.nimbasms.groups;
-
-import gn.nimba.nimbasms.common.Pageable;
-
-public interface Group extends Pageable<GroupResponse> {
-}
