--- src/main/java/gn/nimba/nimbasms/messages/Number.java
@@ -1,12 +1,0 @@
-package gn.nimba.nimbasms.messages;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class Number {
-    private String id;
-    private String contact;
-    private String status;
-}
