--- src/main/java/gn/nimba/nimbasms/sendernames/SenderName.java
@@ -1,6 +1,0 @@
-package gn.nimba.nimbasms.sendernames;
-
-import gn.nimba.nimbasms.common.Pageable;
-
-public interface SenderName extends Pageable<SenderNameResponse> {
-}
