+++ src/test/java/com/nimbasms/nimbasms/NimbaSMSApplicationTests.java
@@ -1,0 +1,6 @@
+package com.nimbasms.nimbasms;
+
+
+class NimbaSMSApplicationTests {
+
+}
