--- src/test/java/gn/nimba/nimbasms/NimbaSMSApplicationTests.java
@@ -1,6 +1,0 @@
-package gn.nimba.nimbasms;
-
-
-class NimbaSMSApplicationTests {
-
-}
